+++ README.md
@@ -1,3 +1,3 @@
-# trace
+# pyhunters
 
 something something
