--- pants.toml
@@ -1,19 +1,0 @@
-[GLOBAL]
-pants_version = "2.25.0"
-backend_packages = [
-    "pants.backend.python",
-    "pants.backend.experimental.python.lint.ruff.check",
-    "pants.backend.experimental.python.lint.ruff.format",
-]
-local_cache = true
-# remote_store_address = "azure/address:port"
-
-[python]
-enable_resolves = true
-interpreter_constraints = ["CPython==3.13.*"]
-
-[source]
-root_patterns = ["/"]
-
-[cli.alias]
-test-all = "test --debug tests::"
