+++ pyhunters/core.py
@@ -4,32 +4,32 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Self
 
-from tracer.mark import Mark
-from tracer.type_mods.singleton import Singleton
+from pyhunters.target import Target
+from pyhunters.type_mods.singleton import Singleton
 
-__all__ = ["mark"]
+__all__ = ["PyHunters"]
 
 
 @dataclass
-class Tracer(metaclass=Singleton):
+class PyHunters(metaclass=Singleton):
     """Class for adding and managing marks."""
 
     name: str
 
     def __post_init__(self):
         """Ensure that the exit handler is registered."""
-        self._marks: list[Mark] = []
+        self._targets: list[Target] = []
         self._register_exit()
 
     def __iter__(self):
         """Iterate through all the marks."""
-        return iter(self._marks)
+        return iter(self._targets)
 
-    def __getitem__(self, key: str) -> Mark:
+    def __getitem__(self, key: str) -> Target:
         """Get a mark by name."""
         return self._map[key]
 
-    def get(self, name: str) -> Mark | None:
+    def get(self, name: str) -> Target | None:
         """Get a mark by name, or None if not found."""
         try:
             return self._map[name]
@@ -37,17 +37,19 @@ class Tracer(metaclass=Singleton):
             return None
 
     def _register_exit(self):
+        """Register the exit handler to summarize all targets."""
         atexit.register(self.summarize)
 
     @property
     def _map(self):
         """Get a dictionary mapping mark names to marks."""
-        return {mark.name: mark for mark in self._marks}
+        return {mark.name: mark for mark in self._targets}
 
-    def add(self, mark: Mark) -> Self:
-        if not isinstance(mark, Mark):
-            raise TypeError(f"Expected Mark, got {type(mark)}.")
-        self._marks += [mark]
+    def add(self, target: Target) -> Self:
+        """Add a target to the collection."""
+        if not isinstance(target, Target):
+            raise TypeError(f"Expected Target, got {type(target)}.")
+        self._targets += [target]
         return self
 
     def summarize(self):
@@ -55,14 +57,14 @@ class Tracer(metaclass=Singleton):
         pass
 
 
-def getTracer(name: str = __name__) -> Tracer:
-    """Create or get the singleton `Tracer` instance."""
-    return Tracer(name=name)
+def getPyHunters(name: str = __name__) -> PyHunters:
+    """Create or get the singleton `PyHunters` instance."""
+    return PyHunters(name=name)
 
 
 def mark(name: str, *, project: str = "DEFAULT"):
     """Simple interface for adding a marker."""
-    tracer = Tracer(project)
+    tracer = PyHunters(project)
     caller = inspect.stack()[1]
     module_path = Path(caller.filename)
     line_no = caller.lineno + 1
@@ -82,10 +84,10 @@ def mark(name: str, *, project: str = "DEFAULT"):
                 returns = func(*args, **kwargs)
             except Exception as exc:
                 mark_kwargs["error"] = exc
-                tracer.add(Mark(**mark_kwargs))
+                tracer.add(Target(**mark_kwargs))
                 raise exc
             mark_kwargs["returns"] = returns
-            tracer.add(Mark.model_validate(**mark_kwargs))
+            tracer.add(Target.model_validate(**mark_kwargs))
             return returns
 
         return wrapper
