--- terraform/pants.tfvars
@@ -1,1 +1,0 @@
-resource_group_name = "test-resource-group"
