--- terraform/variables.tf
@@ -1,4 +1,0 @@
-variable "resource_group_name" {
-  type        = string
-  description = "The name of the resource group"
-}
